Fix tests refactor part 1
