+++ app/src/androidTest/java/com/github/polybooks/activities/MainTest.kt
@@ -7,6 +7,7 @@ import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
+import com.github.polybooks.LoginActivity
 import com.github.polybooks.R
 
 import org.hamcrest.Matchers.not
