+++ app/src/main/java/com/github/polybooks/activities/MainActivity.kt
@@ -5,8 +5,8 @@ import android.os.Bundle
 import android.view.View
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
-import com.github.polybooks.LoginActivity
 import com.github.polybooks.R
+import com.github.polybooks.activities.LoginActivity
 import com.github.polybooks.utils.setupNavbar
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.google.firebase.auth.FirebaseAuth
